feat(flotilla): Fault tolerance (#4628)
## Changes Made

Build fault tolerance into flotilla.

Instead of using ray retries on the actor, we catch the actor error when
awaiting the results. If the actor died, is unavailable, or
unschedulable, we propagate this information back to the scheduler.

The scheduler will mark that worker dead and retry the task on the
separate worker. If there was only ever 1 worker and it died, its ok
because on every scheduler tick we will try to get new workers.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)