feat(flotilla): support task progress for logical plan graph (#4686)
## Changes Made

This PR adds support for reporting progress on a logical graph based on
currently executing tasks:
- In
[`run_plan`](https://github.com/Eventual-Inc/Daft/blob/0398fe08b5b92c1ee055fe1ac31b1c848adeb542/src/daft-distributed/src/plan/runner.rs#L65-L66)
we receive a `DistributedPhysicalPlan` which now stores a reference to
the _optimized_ logical plan that it was translated from.
- All logical nodes are assigned an ID but only after the logical plan
has been optimized. This ID is forwarded to physical nodes so that tasks
executing based on the physical plan can be correlated back to the
logical plan that produced them.
- Introduces a new `HttpSubscriber` which aggregates plan and task
events, reporting a query graph structure to the dashboard URL
configured in environment variables. The query graph structure mirrors
the logical plan graph structure since this is semantically intelligible
to the user and allows us to obscure the complexity of Flotilla
internals. It defines contract between the dashboard and the engine,
which will allow the engine to render additional node-specific metadata
without requiring constant dashboard updates.

Next steps:
1. Update the OSS dashboard to accept this query graph structure
2. Support common and custom metrics for each physical node type
3. Improve HttpSubscriber—we should avoid locking on the main thread and
should also batch task updates to be sent every 1 second, rather than
making a request on every task update.

I'd like to merge this in its current state and address the above next
steps in smaller followup PRs since the new query observability
functionality is opt-in and there should be no impact on performance
when it is disabled, which is the default behavior.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->
N/A

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)

---------

Co-authored-by: Colin Ho <colin.ho99@gmail.com>