chore: add DAFT_LOG env indicates log level (#4807)
## Changes Made

<!-- Describe what changes were made and why. Include implementation
details if necessary. -->

Currently, if you want to set the log level, you need to follow the
following method, which is rather cumbersome. Therefore, an environment
variable is introduced here as the top - priority identifier for the log
level of Rust.
```
logging.basicConfig(
    level=logging.DEBUG,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    stream=sys.stdout
)
daft.refresh_logger()
```

In addition, some instructions on how to view debug logs in pytest were
added to CONTRIBUTE.md.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)

---------

Co-authored-by: Colin Ho <chiuhong@usc.edu>
Co-authored-by: Colin Ho <colin.ho99@gmail.com>