fix: update a new expression example in the contributing guide to work (#4818)
## Changes Made

A new expression example in CONTRIBUTING.md needed couple of small
changes to work. Current implementations of the expressions in the code
base is more succinct than this example but the example is simpler to
read so instead of replacing it with the newer code (similar to what's
used in the rest of the code) I just fixed it to work.

## Related Issues

No related issues

## Checklist

There are no changes to docs

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)

---------

Signed-off-by: Niko Ulmasov <ulmasov@hotmail.com>