chore: upgrade py-spy version (#4830)
## Changes Made

Because the current version of py-spy in ray is `py-spy==0.4.0`, while
the version in daft is `py-spy==0.3.14`. Introducing them simultaneously
would cause a version conflict. Therefore, a small change was made here
to change the version of py-spy to a range version.

<!-- Describe what changes were made and why. Include implementation
details if necessary. -->

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)