feat: add support for reading json arrays (#4844)
## Changes Made

General approach is that we peek at the first byte of the file. if it's
`[` we use the array algorithm, if its `{` we use the default ndjson
algo.


⚠️ 

It's worth noting that this is not highly optimized and _could_ OOM on
very large JSON files as they are pulled **entirely** into memory.


as a side note: 
Our JSON reading logic is a total mess and I didn't want to do any
unnecessary refactors, so there was a bit of copy/paste in this PR to
get things to work.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)