feat: Support multiple namespaces in the turbopuffer writer (#4845)
## Changes Made

When writing to turbopuffer, it's common to write data to multiple
namespaces. We extend the turbopuffer data sink so that the `namespace`
parameter can take either a string or an expression. If we're given a
string, all data gets written to the given namespace. Otherwise, we
evaluate the expression and partition the data by namespace value before
writing to their respective namespaces.