chore: new code path for scalar UDF (#4814)
## Changes Made

Was trying to get kwargs to work with scalar UDFs, decided to bite the
bullet and build out the entire code path for the new UDFs so that it no
longer relies on the legacy UDF code. Turned out to be quite simple
actually, the complexity of the legacy UDF code is largely from the
class UDF stuff we did.

Additional changes:
- add ability to specify kwargs to UDF
- remove `@daft.func.batch`. It was just an alias to `@daft.udf` anyway,
I think we should actually think about it a bit more since I don't think
I want the batch UDF to behave exactly like the current UDF either.
- Rename Python UDF to legacy Python UDF

## Related Issues

Closes https://github.com/Eventual-Inc/Daft/issues/4729

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)