fix: Fix list sorting when operating on a morsel slice (#4863)
## Changes Made

When we split data into morsels, list operations operate on a slice of
the original list. In this case, when performing a list sort, we create
new lists that should have offsets starting from 0. However we
incorrectly assumed that offsets do not change when sorting. This PR
fixes that.

## Related Issues

Related to #4862. We need a proper audit to close the issue.