feat: Interactive jupyter display (#4835)
## Changes Made

When user clicks on a cell, it opens a side pane that shows the full
content for that cell. Currently works for images and strings only.

This works by running an http server that serves the cell content, and
embedding javascript logic in the html to make these fetches.

**Error behavior:**
- If the server is not running / cannot connect to it, the cells will
not be clickable.
- If the server was initially running and then stopped, the cells will
still be clickable and the sidepane will still open, but the original
truncated content will be served.

**Notes:**
- Some kind of server is necessary to serve the content because we
embedding the full content in the html in the beginning can be too
heavy, affecting latency and potentially causing crashes.
- The alternative to server was ipywidgets or ipykernel comm.
- pywidgets will introduce extra dependency, and they are also quite
opaque, hard to debug / optimize.
  - ipykernel comm does not work on all jupyter environments.
- I leveraged the dashboard for the http server, because we already have
it. Plus it can potentially save some work for when we do dataframe
displays in the dashboard.

This PR was tested on cursor jupyter, jupyter notebook, and jupyter lab.

Generated a bunch of 4k images and large strings to test:


https://github.com/user-attachments/assets/8ad5c677-ea35-471a-ac29-c358dd02e468




## Related Issues

Closes https://github.com/Eventual-Inc/Daft/issues/4396

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)