chore: change series literal to list (#4896)
## Changes Made

Remove `LiteralValue::Series` and replace it with a
`LiteralValue::List`. I don't think we use series literals as series
anymore so this name makes more sense. Additionally I've changed the
datatypes/conversions to reflect that this is actually a list-type
scalar value.

This is part of a larger change that I'm working on to unify type
conversions between Python and Daft.

Notes:
- Removed `test_length_mismatch_all_types` because length mismatch is
now handled by the RecordBatch constructor so there's a different error,
which should happen for any expression, so no need to test it
specifically for this one.
- Removed `test_hash_int_array_with_bad_length` - same reason as above.
- Removed `test_table_expr_is_in_different_input_types`. It tested
behavior that we did not document, so I'm removing it instead of fixing
it.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)

---------

Co-authored-by: Colin Ho <colin.ho99@gmail.com>