docs: getdaft.io → daft.ai (#4926)
## Changes Made

Some hyperlinks in documentation are still getdaft.io

We used to have a `core_concepts.md` but after reorganizing the docs, we
no longer have that page. For those that I couldn't find a proper
replacement for the link, I opted to just delete the link. LMK if you
have any thoughts or suggestions for where to link things.

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)