feat: adds column set visitor and use in pushdowns (#4929)
## Changes Made

* Adds a visitor which collects a column set
* Uses this in Pushdowns to return the filter column set.

## Related Issues

I considered saving the list from the rust pushdowns object, but a
python visitor is more generic and flexible because it can be used
outside the context of pushdowns to find all the columns in an arbitrary
expression. This also uses a set for unique column names. If/when daft
has a notion of bound columns, then the hash will be important for
distinctness because for now we rely on unique names within an
expression tree.

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)