refactor: move literal to daft-core (#4928)
## Changes Made

This is part of my overall work to [unify type
conversions](https://www.notion.so/Daft-Python-Type-Conversions-240091bd8634808abf18e00abc9b7fc8?source=copy_link)
and is in particular cleanup that I wanted to get done before adding
more literal variants.

The crux of this PR is the move of `LiteralValue` from `daft-dsl` to
`daft-core`. This serves several purposes:
1. Series -> Literal conversion can now be delegated to the individual
array types. This is a lot cleaner than doing manual downcasts based on
dtype, and will allow for conversion into literal to work even if a data
type has multiple physical implementations.
2. `Vec<Literal>` -> Series conversion can now be implemented as a
`TryFrom`, which is more idiomatic. Literal -> Series can also be a
`From` now.
3. This creates a cleaner separation between literals and expressions,
whereas before they could depend on each other. Instead, literals and
series are more tightly associated now, which is more sensible as
literals are pretty much the scalar counterparts to series.

Additional changes:
- Renamed `LiteralValue` to `Literal` and the original `Literal` trait
was removed in favor of implementing `From<Literal>`
- Removed `Literal::FixedSizeBinary`. I don't plan on adding the fixed
size/shape variants for the other types so this was kind of an odd one
out. Removing it does not remove any user-facing functionality anyway,
but it does simplify the implementation so I chose to go with it.
- Added `Literal::Float32`
- Moved `PyObjectWrapper` from `daft-dsl` to `common-py-serde` so we can
use it in `daft-core`

## Related Issues


https://www.notion.so/Daft-Python-Type-Conversions-240091bd8634808abf18e00abc9b7fc8?source=copy_link

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)