perf(embed_text): Let Sentence Transformers select the best available device (#5082)
## Changes Made

From the docs:
> The PyTorch backend is the default backend for Sentence Transformers.
If you don’t specify a device, it will use the strongest available
option across “cuda”, “mps”, and “cpu”.

Instead of setting cuda or cpu, we should let mps be used too so that we
use Metal on MacOS with GPUs.

Planning to explore ONNX and OpenVINO backends next.