feat: Add Arrow IPC conversion for RecordBatches (#5143)
## Changes Made

I need it for both a process UDF PR and for the dashboard stuff, so just
moved it to this PR to rebase the others on top of it. Lmk if you want
me to add a test for this specifically, but I do test it E2E in the
others.