ci: remove macos from PR test suite (#5142)
We currently use MacOS to run unit tests in PRs because we had an issue
with getting Rust code coverage in Ubuntu 22.04. The latest Rust
toolchain no longer has this issue, so this PR updates the rust
toolchain to the latest nightly and moves the code coverage tests to
Ubuntu.

The vast majority of the diff here is just linting and lifetime fixes
due to the upgrade to the latest nightly Rust + 2024 edition. The rest
of the changes are as follows:
- moving the code coverage tests in
`.github/workflows/pr-test-suite.yml` to Ubuntu as described above
- it was failing due to out of disk space so I also added a disk space
remover action
- update our Rust versions in `rust-toolchain.toml` and `Cargo.toml`
- use `hashbrown::HashMap` instead of `std::HashMap` because
`raw_entry_mut` was removed from the standard library

https://github.com/Eventual-Inc/Daft/issues/3801

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)