ci: fix TPC-H benchmark workflows (#5123)
## Changes Made

Fixes the UV setup in the benchmark workflows so that they work with our
new dependency setup.

I was having issues with running a Daft script as a Ray job and it turns
out that we broke the behavior of automatically connecting to the Ray
cluster when running in a Ray job in #4567.

The `RAY_ADDRESS` env var is set to `<local ip>:6379` in a job script,
but in that PR we prepended `ray://` to it, which causes `ray.init` to
try and fail to create a Ray client instead of connecting to the
cluster. I've removed that logic, as well as the logic in Daft to read
`RAY_ADDRESS` because `ray.init` does that automatically already.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)