fix: remove io_config from global scan operator (#4676)
## Changes Made
The configuration items of IO_Config are displayed in GlobScanOperator,
but not in physicalScanOperator. This PR attempts to unify this behavior
by also not showing IOConfig in GlobScanOperator. Additionally, items
like s3 and gcs in IOConfig should all be optional. They are optional in
the Python interface, but have become mandatory in Rust. So here, they
are made optional, and only the configured items will be displayed.

* old

![image](https://github.com/user-attachments/assets/1744fd50-916a-43f5-bc5f-6d9e612204d0)

* new

![image](https://github.com/user-attachments/assets/ca3e9a4a-36c4-4a8f-9d6a-fa99ae1cbd1b)

<!-- Describe what changes were made and why. Include implementation
details if necessary. -->

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)