perf: Drop local limit tasks global limit reached (#5194)
## Changes Made

Right now we wait for all the local limit tasks to complete even if we
have received enough, which is unnecessary. Sometimes the output of the
`.show` will be displayed, and then the progress bar keeps going for a
long time because tasks are waiting to be finished, which is very weird.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)