fix: Cast to table schema in lance write (#5221)
## Changes Made

If the dataset has a schema, we can cast the data to that schema before
writing. This will solve the case where dataset has a string column, but
the daft table has a large string column (by default the daft utf8 dtype
is converted to pyarrow large string).

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)