feat: allow PythonArray to be serialized (#5270)
## Changes Made

Reimplemented `PythonArray` so that it's serializable via both serde and
to/from arrow. Removed the `PseudoArrowArray` stuff so now `PythonArray`
does not hold any sort of arrow2 array.

One breaking change is that `agg_list` on a Python column now produces a
`List[Python]` column since that is now serializable. I'm not
particularly concerned about this though because it never really made
any sense for it to aggregate into a Python type.

## Related Issues

Blocking  #5201

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)

---------

Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>