test: Complete coverage for array comparisons (#5286)
## Changes Made

Based on codecov, our array comparison code has 491 lines of untested
code:
https://app.codecov.io/gh/Eventual-Inc/Daft/tree/main/src%2Fdaft-core%2Fsrc%2Farray%2Fops

I was curious to see how good codex is so I gave it a prompt to add
tests for the misses. It did a remarkable job planning out all the tests
we needed, but didn't write compilable code.

I then let Cursor with claude-4-sonnet in agent mode try to fix it. Not
bad, and we actually caught a missing implementation error!

Alas, claude-4-sonnet simply added a `#[should_panic]` tag to the test
case. Bad claude.

Made a small fix for that case.