perf: more literal optimizations (#5314)
supercedes #5305  
## Changes Made

### Adds new method `Series::from_literal_iter`.
This allows us to create series from an iterator of
`DaftResult<Literal>` so we don't need to force a materialization
beforehand.

- The existing `Series::from_literals` is unchanged, but uses this new
iter method internally.

- Some of the more complex types still require a materialization, but
many can be lazily created.

- the `from_literal_iter` also has a new parameter to indicate how to
handle errors. Right now it's always set to `Raise`, but the idea here
is that for udf's users will likely want a way to just ignore errors and
set them to null. This was added because I needed to handle result types
anyways. So seemed like a nice addition.

- `from_literal_iter` also skips dtype inference and is expected that
the literals match the variant of the input dtype.



### Series literal conversion optimizations
- use MutablePrimitiveArray when possible instead of `T::from_iter(...)`
- remove some extra bounds checks inside `get_lit`. This was directly
calling `.get` which was also doing bounds checks. So we were checking
the bounds twice.



## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly