feat: @daft.cls decorator for new class UDFs (#5350)
## Changes Made

Add the `@daft.cls` and `@daft.method` decorators for constructing a
Daft class.

Some additional changes:
- Did a bit of refactoring on the Python side because I did not want to
have to implement a whole matrix of UDFs (func vs class, row-wise vs
async vs generator). Ended up combining all the function variants into a
`Func` class and the code is decently clean.
- Looks like we added `use_process` in #5323 but I don't think it was
actually hooked up to use processes under the hood. My changes in this
PR should fix that
- Restructured the class-type decorators since overriding `__new__`
seems to make type checkers unhappy. Resolves #5284

I initially implemented a `@daft.setup` that users would use in lieu of
having an `__init__` method, but after writing several examples with it,
it just felt like a very obtuse way to introduce this whole new thing
just to tell users that initialization is lazy. We'd have to explain
that the arguments in the constructor are passed into the setup method,
and that you have to add this decorator. We had some discussion about
this in the working session and I don't think others thought a lazy init
would be an issue anyway.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly