feat: pydantic model type conversion (#5370)
## Changes Made

Add ability to convert pydantic models to Daft struct as well as type
inference for pydantic models.

Also adds `packaging` as a default dependency because we use it in a
variety of places now. For example our `to_pandas` uses it as well, but
we do not have a Pandas requirement that includes `packaging`. Based on
the latest versions, the package itself is 66.5 kB, whereas Daft is
between 43.0 and 47.2 MB, so this will only result in a 0.15% increase
in download size at most.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly