feat: Explicit AWS vs. HTTP mode for common crawl dataset (#5379)
Adds a new required argument to `daft.datasets.common_crawl`: `in_aws:
bool`. This **must** be set to `True` when running in AWS and `False`
when running outside of AWS. This allows Daft to select the most optimal
download strategy for CC data. Added a notice about this to the
docstring.

Refactors the existing mocked unit tests for this by making the tests
patch the appropriate `_get_{s3,http}_manifest_path` using the value of
`in_aws`. Adds `in_aws` as a pytest parameter and parameterizes each
test on `True` and `False`.

Updates the Common Crawl documentation to mention the new required
`in_aws` parameter. Adds a new section discussing the new HTTP
download mode and provides an example.