feat: add mimetype detection for daft.file (#5411)
## Changes Made

adds mimetype detection for daft.file. 

If it's url based, will first attempt to look at the file extension, If
that is unavailable, it will look at the first 16 bytes to try to
determine the mimetype.

If the mimetype is undetectable, the python api will return the default
`application/octet-stream`.



## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly