feat: chat completions api for prompt function (#5497)
## Changes Made

Add ability for users to choose chat completions instead of responses
when prompting with openai provider.

As a drive by this PR also fixes the file input calls which were
previously failing due to incorrect message schema as well as non pdf
files. Support for text files that are non-pdf can be achieved by
injecting the contents into the text input, which will be done in a
subsequent pr.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly