feat: emit selectivity metric to OTel in swordfish filter op (#5584)
## Changes Made

Adds an OTel Gauge for the selectivity metric that the swordfish filter
op calculates, in order to emit it to OTel as well as to the internal
stats.

This recalculates selectivity on every row in or out, rather than the
existing approach of only calculating it at snapshot time, since the
Gauge could be polled at any time for the current value. It's
theoretically possible to recover the previous semantics by using an
`ObservableGauge` which uses a callback to calculate the gauge value on
request, but that gets tricky with lifetimes.

This also stores our own record of the selectivity in an `AtomicF64`
from the `atomic_float` crate, which does [unsafe
shenanigans](https://github.com/thomcc/atomic_float/blob/main/src/atomic_f64.rs#L51-L53)
to store an f64 in an `AtomicU64`. It seems to be fairly well thought
out so I think it's fine.

## Related Issues

None.

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly