fix: Fix blocked and oom issues for scan lance (#5592)
## Changes Made

<!-- Describe what changes were made and why. Include implementation
details if necessary. -->

When building Scan Tasks, avoid directly binding the lance dataset
object to each Scan Task. And the size of a single lance dataset is
approximately 0.76MB.
`self._ds: 0.76 MB`

First, split the attributes of the lance dataset. The Scan Task only
binds the attributes of the lance dataset. When it is actually executed,
refactor it back to the lance dataset to reduce the size of the
execution plan.

For such an execution plan containing 8,354 scantasks, the size of the
optimized execution plan has been reduced from over 13GB to 25MB.

The execution time ranges from execution failure to returning results
within seconds.

```
== Physical Plan ==

* Limit: 10
|
* ScanTaskSource:
**|   Num Scan Tasks = 8354**
|   Estimated Scan Bytes = 0
|   Pushdowns: {limit: 10}
|   Schema: {seq_id#Int64, ext#Utf8, file_size_in_byte#Int64, program_lang#Utf8,
|     doc_type#Utf8, text#Utf8, filename#Utf8, repo_name#Utf8, sub_path#Utf8,
|     stars#Utf8, dataset#Utf8, pt#Utf8, tar_path#Utf8, file_path#Utf8}
|   Scan Tasks: [
|   {daft.io.lance.lance_scan:_lancedb_table_factory_function}
|   {daft.io.lance.lance_scan:_lancedb_table_factory_function}
|   {daft.io.lance.lance_scan:_lancedb_table_factory_function}
|   ...
|   {daft.io.lance.lance_scan:_lancedb_table_factory_function}
|   {daft.io.lance.lance_scan:_lancedb_table_factory_function}
|   {daft.io.lance.lance_scan:_lancedb_table_factory_function}
|   ]

```

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

https://github.com/Eventual-Inc/Daft/issues/5591

## How to test
existing UTs.

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly

---------

Co-authored-by: 蔡灿 <caican@xiaomi.com>