feat: Streaming sample by size (#5663)
## Changes Made

Implement a streaming sample by size algorithm for swordfish.
- For sampling without replacement, we assign a random number to each
row, and simply keep around the rows with the lowest N random numbers.
- For sampling with replacement where size=N, we create N slots. For
each slot, assign a random number to each row. If the random number is
smaller than the existing one in that slot, replace it.

This PR only implements for swordfish. For flotilla, we will need to do
some merging of local samples, this will be in the next PR. Essentially,
we keep the random numbers per row / slot, merge the data onto a single
node, and then pick the N rows with the smallest random numbers (per
slot if with replacement).

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->